Create README.md
stellar veiws: NASA Picture of the Day.

Welcome to the NASA Picture of the Day repository! This website features stunning images captured by NASA's satellites, telescopes, and other equipment. Every day, a new picture is uploaded to showcase the beauty of our universe and the incredible work done by NASA scientists and engineers.
Usage

To view the Picture of the Day, simply visit the website at [INSERT WEBSITE URL HERE], and the latest image will be displayed. You can also browse through our archives to view previous pictures of the day, dating back several years.
Contributing

We welcome contributions from everyone! If you have a suggestion for a future Picture of the Day or would like to contribute code to improve the website, please feel free to submit a pull request.

To get started, you will need to have Git installed on your local machine. Fork this repository to your own account and then clone it to your machine. Make sure to create a new branch for your changes.